+++ examples/displacement_mapping.rs
@@ -224,7 +224,7 @@ fn main() {
                     }),
                     &program, &uniforms,
                     &Default::default()).unwrap();
-        target.finish();
+        target.finish().unwrap();
 
         // polling and handling the events received by the window
         for event in display.poll_events() {