+++ examples/tutorial-03a.rs
@@ -70,7 +70,7 @@ fn main() {
 
         target.draw(&vertex_buffer, &indices, &program, &uniforms,
                     &Default::default()).unwrap();
-        target.finish();
+        target.finish().unwrap();
 
         for ev in display.poll_events() {
             match ev {