+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## Unreleased
 
+ - `Frame::finish` now returns a `Result`. `Frame`'s destrutor will panic if `finish` has not been called.
  - `with_compressed_data` and `with_compressed_data_if_supported` now have an additional parameter for mipmaps.
  - All the texture constructors that used to take a boolean as parameter for mipmaps now takes an enum.
  - `empty_with_format` and `empty_with_format_if_supported` are now allowed for compressed textures.
  - `write`, `write_compressed_data`, `write_compressed_data_if_supported` and `read_compressed_data` are now available for mipmap objects.
  - Removed the `is_closed()` function.
  - Removed the deprecated `render_buffer` module.
+ - The `Backend::swap_buffers` function must now return a `Result`.
 
 ## Version 0.5.6